Remove the empty nav2_gazebo_localizer directory (#155)
Instead, we can add the fake_localization package to the repo and
finish the port to ROS2. This will allow us to test the stack with
a stub localizer, thereby avoiding the ros1_bridge when running
in simulation, since AMCL currently needs input from gazebo which
must be run on the ROS1 side since since the ROS2 version of the gazebo
plugins are not yet available.