Adding timeouts to AMCL transform calls (#166)
This is a a workaround. The current AMCL code frequently fails to
transform data and generate a pose. The current hypothesis is that
this is a result of the message filter code having been removed.

This change works around the problem by adding a 1 second timeout
on each transform, meaning there is now a 1 second window from
when the data is received till the transform can become available.

It's not clear whether this workaround has any unintended
consequences. Specfically I'm concerned about performance and
latency.