Address issue #164 whee the goal tolerance is hardcoded in the SimpleNavigator
Instead, the tolerance should be provided along with the goal pose. The intention
being that the command would be: "navigate to this specified goal pose within
this specified tolerance." So, I created a new message, PoseWithToleranceStamped,
to hold a pose with an associated tolerance. This is now used as the command
message to NavigateToPose. I've also updated the command message to ComputePathToPose,
which is a PathEndPoints message to use this new message type. As a result the
PathEndPoints message/command has the current pose and the goal pose w/ tolerance.