Get rid of the Robot abstract base class (#234)
Since we likely won't implement non-ROS robots, remove the abstract base class
and make the Robot class be a ROS robot. The Robot class will contain the functionality
common to all ROS robots. There can be derived classes that add functionality that is
not common to all ROS robots.

Also, since I'm modifying the SimpleNavigator, update the ROS log messages so that
they don't (redundantly) include the module name since it's already in the output.