Consistent package format
Some of the packages used http://download.ros.org/schema/package_format2.xsd
that requires package format to be set to `2`, but package format was
set to 3 instead.

This change makes them consistent.  As a result linter for packages
won't compain on this error.  Version 3 picked up because it is superset of
version 2 with addition of optional fields.