Rename the costmap_world_model
The original idea was that we might have different world model implementations.
However, as it turns out, we will likely only have the one implementation that
evolves over time. So, we don't necessarily want to emphasize that it is
costmap-based, but want to allow for other representations going forward. Hence
the name change.