Ignore Visual Studio Code files
Using Visual Studio Code, I've been opening just the directories that
I'm interested in, avoiding the output directories (install, build,
and log). In this configuration, when debugging various modules, the
VS Code app will put .vscode files in each source directory being
debugged. This causes the directories to be highlighted in the editor
indicating that files have been added. With this addition to .gitignore,
these files will be ignored and the VS Code editor will more accurately
reflect any changes to the workspace.