Fix issue #44 to modify TaskServer to accept a ROS node rather than be a ROS node
Updated the TaskClient and TaskServer templates to accept a rclcpp::Node::SharedPtr
reference. Also updated the Robot class to do the same. Also had to then change
the execute method to be set on the TaskServer rather than have derived classes
override the method. Now, it is more like the SimpleActionServer from ROS1,
which sets us up a bit better for the port to Actions when ready on ROS2.