Remove costmap movement callback (#303)
* Disabling the costmap movement callback.

This callback produces lots of error message noise. It also has no
actual function and should be removed. It sets a couple of variables
on the costmap object, but nothing actually reads those variables.

* Removing movement callback from costmap_2d_ros

It had no effect. It set variables that were unused in the rest
of the system.