Removing the validCellToQueue optimization from MapGridQueue.
The costmap_queue concept as implemented is incompatible with not
queuing some cells. The costmap_queue expands adjacent cells as
the least cost cell is pulled from the queue. For it to work
properly, the newly added cells must have a cost that is greater
than or equal to the current cell.

This requirement is violated in map grid when there are obstacles.
The cells in the shadow of the obstacle have a lower cost than those
outside the shadow, but they don't get expanded till a cell outside
the shadow is expanded.

If a lower cost cell is added, the costmap_queue iterator does not
go backwards, so those lower cost cells never get pulled from the
queue. If the queue iterator were instead "fixed" to go backwards,
it would no longer guarantee cells were pulled from the queue in
order of cost, violating the fundamental concept of that structure.

A possible fix would be to fully expand the map grid before
trying to pull any cells from the queue. I need to explore that
possibility for performance issues.