Add node name filter to ServiceClient.
ServiceClient reuses the service name as the temporary node name.
However, service names frequently have / characters in them and
that is invalid for a node name. Also, it seems likely we'll have
multiple clients using the same service. This would result in a name
collision. This now adds a random suffix to the internal node name.