Remove the nav2_mission_executor directory
This functionality will be removed from the nav2 repository and eventually
revived in a separate repository. The plan is to separate out the generic
Behavior Tree functionality first into a separate repo and then we can
build a mission execution system using this generic BT functionality,
also in a separate repo. Then, we will be able to span navigation,
grasping, and other kinds of tasks, all coordinated using BTs.