Delete pause/resume buttons and services
We don't yet have a full solution for bringing nodes into a state where
their parameters can be reconfigured and then the nodes reactivated. So,
we don't want to expose a partial solution for the Dashing release that
could potentiall be confusing for users.

Currently, all lifecycle node parameters are read during on_configure.
Going forward, we will re-assess all parameters to allow for other
options, such as on_activate. Once parameters are associated with different
state transitions, we'll then have to clearly document the parameters and
their state associations and/or come up with an automated way to allow for
node reconfiguration.