Call now on the main lifecycle nodes instead of rclcpp_nodes (#845)
Now that lifecycle nodes properly support use_sim_time, we can get the
current time from the node itself rather than relying on the rclcpp_node,
which is a temporary bridging solution to interface to ROS2 functionality
that doesn't yet support lifecycle nodes.  This PR is one of several steps
that we'll take in order to remove the rclcpp_nodes from the navigation
code, thereby reducing the overall number of nodes in the system and
therefore the overall number of topics and services.