Consolidate related parameter sections under a single namespace heading
The navigation2 launch implements rewriting of the input parameter file in
order to change some values, such as use_sim_time, based on command line
input. This is currently required to work around some node renaming that
occurs if we use the Node class in the launch system.

The input nav2_params.yaml file had separate sections for nodes under
the global_costmap and local costmap namespaces. This caused some issues
for our parameter rewriting code (see attached files). So, to work around
this problem and to use what is probably a better organization anyways,
this PR puts the nodes under the same namespace section.