Switch from map service to topic (#885)
* Add parameter to disable map server

* Move parameter check to getParameters

* Uncrustify

* Don't mark static layer as current until we have a map

* Make sure static layer gets initialized

Make sure static layer gets initialized, even with an empty layer, so it doesn't crash if you disable the map service and enable the obstacle layer

* remove map service from static layer

Only use map topic instead
Subscribe to map topic with default transient local volatility
Add parameter "map_subscribe_transient_local" (default=true) for backwards compatibility with volatile publishers

* Change AMCL to use only the map topic, not map service

* remove now-unused MapServiceClient

* uncrustify

* only set current_=true when first map received

* Fix nav2_costmap inflation tests.

* Fixing collision tester test.

* Revert "uncrustify"

This reverts commit 82f88a73006d8c696625c1ce8327659654745627.
This uncrustify failure is enforced differently depending on whether
we are on the dashing branch or on the master branch. This particular
commit is only appropriate on the dashing branch, but is already present
there.