Never shortcut trajectory scoring
When evaluating trajectories, dwb will stop running critics as soon as
the trajectory's score is better than the best trajectory seen so far.
It looks like this often leads to worse trajectories being chosen as the
overall best.

Consider 3 trajectories and 2 critics:

1. t1 is scored by both critics and gets scores [0.5, 0.0] for a total
score of 0.5
1. t2 is scored by the first critic and gets a score of 0.6. Since
it's better than the best so far, the scoring gets cut short. If it
hadn't been cut short, t2's scores would have been [0.6, 0.5] for a
total of 1.1
2. t3 is scored by both critics for scores [0.4, 0.4] for a total of
0.8. t3 is chosen as the best trajectory even though it's total score of
0.8 is smaller than what t2's would have been if all critics had been
run on it.

Signed-off-by: Shane Loretz<sloretz@openrobotics.org>