Add a min_obstacle_height param to the nav2_costmap_2d's ObstacleLayer plugin (#3211)
This allows considering full range observations, specified by the
  <data source>.min_obstacle_height
and
  <data source>.max_obstacle_height
especially used for the raytracing, but to still be able to specify a
minimum obstacle height to report obstacles onto the costmap.

This is in particular required in the case a PointCloud2 source points
slightly towards the ground, sometimes detecting obstacles, that should be
cleared once the ground reappears behind the obstacle when it has moved
away: we don't want to detect the ground as an obstacle, but still want it
to be used in the raytracing to clear the previously detected obstacle.

(cherry picked from commit 0b4179b6432c62efdf1e853bce382de2e6706061)