nav2_dwb_controller: enhance path pruning and shortening mechanisms
The pruning and shortening mechanisms allow to determine a subset of the
path to be passed to the critics. Those mechanisms ignored every point
of the path which euclidian distance to the robot was above a defined
threshold (prune_distance and shorten_distance). As a consequence, those
mechanisms were rejecting more or less points depending on the robot
transversal distance from the path. When this distance was
non-negligible w.r.t. the thresholds, the remaining points might not
be sufficient for an effective path tracking and the robot could get
stuck. Even worse, when the robot was further than prune_distance away
from the path, the latter was progressively erased at a rate of 1 point
per iteration. This issue prevented from setting the prune distance to
a small value close to zero to instantaneously remove the passed points.

The proposed mechanism iterates through the path until finding the
closest point to the robot. From there, it goes through the path
in both direction and measures the integrated distance until reaching
prune_distance for passed points and shorten_distance for upcoming
points.