Publish collision monitor state (#3504)
This commits is cherry-picked from b8d077e. The collision monitor now
has an additional parameter named "state_topic". If the state topic is
set, the collision monitor will publish its status to the defined topic
so that other nodes might perform actions based on the current state of
the collision monitor.

Signed-off-by: agennart <antoine.gennart@quimesis.be>