Revamp nav2_util CMakeLists.txt to use modern idioms.
This commit does a number of things:

1.  Switches to using target_link_libraries everywhere.
This gives us finer-grained control over what dependencies
are exported to downstream as public, or private.  In the
particular case of nav2_util, this actually doesn't matter
*too* much, but it will help for other packages.
2.  Moves the include directory down one level to
include/${PROJECT_NAME}, which is best practice in ROS 2
since Humble.
3.  Removes the separate src/CMakeLists.txt.  While this
isn't strictly necessary, it makes it a lot easier to do
the rest of the changes in here.
4.  Makes sure to export nav2_util as a CMake target, so
downstream users of it can use that target.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>