Revamp nav2_behavior_tree CMakeLists.txt to use modern idioms. (#4485)
This commit does a number of things:

1.  Switches to using target_link_libraries everywhere.
This gives us finer-grained control over what dependencies
are exported to downstream as public, or private.  In the
particular case of nav2_behavior_tree, this actually doesn't matter
*too* much, but it will help for other packages.
2.  Moves the include directory down one level to
include/${PROJECT_NAME}, which is best practice in ROS 2
since Humble.
3.  Makes sure to export nav2_behavior_tree as a CMake target, so
downstream users of it can use that target.
4.  Removes the use of boost.  To do this, we had to introduce
our own version of the "split_string" method.
5.  Moves the test_action_server.hpp file into the main include
directory.  This is slightly odd, but because downstream packages
(opennav_docking_bt) depend on this header file to compile
their own tests, it is technically part of the public interface.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>