Switch nav2_amcl to use modern CMake idioms.
This change is a little unusual in that it does rearrange
the hierarchy of CMakeLists.txt files.  This is because
the individual libraries (like map_lib) are used in some of
the other libraries, and CMake was complaining about that.
So I flattened the whole thing out and now build all of the
libraries at the top-level.

This PR also contains a couple of fixes to previous packages
where I failed to properly export the include directory.

Finally, this PR goes ahead and removes quite a bit of dead
code from this package.  This has been dead for quite a while,
so I think it makes sense to remove it at this point.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>