Make nav2_core an INTERFACE library.
That is, it exports header files, but it does not have
any implementation itself.  Make sure to make an INTERFACE
library so downstreams can have access to the header files.

Note that this also required a change to nav2_graceful_controller,
since it was erroneously getting a visualization_msgs dependency
from nav2_core.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>