nav2_map_server: map_io: generalize computation of raw map pixel value
This implementation is inspired from the Costmap2D implementation [1].
It remains valid even if the LETHAL_OBSTACLE, FREE_SPACE,
OCC_GRID_OCCUPIED, and OCC_GRID_FREE constants are modified in the
future. Unfortunately, the nav2_costmap_2d namespace cannot be accessed
from this module, so the LETHAL_OBSTACLE and FREE_SPACE constants had
to be hard-coded.

By the same way, this commit fixes an issue in the output image where
the areas of NO_INFORMATION and LETHAL_OBSTACLES cannot be distinguished
as they both use the same value: 255. Now, the LETHAL_OBSTACLES pixels
are restored to their rightful value: 254.

[1] https://github.com/ros-navigation/navigation2/blob/main/nav2_costmap_2d/src/costmap_2d.cpp#L80