[RotationShimController] fix: rotate on short paths
Add header data to goal for short paths.

Commit d8ae3c1f9b8233e86ed54dfbe615b1ba56b51b6d added the possibility to
the rotation shim controller to rotate towards the goal when the goal
was closer that the `forward_sampling_distance`. This feature was not
fully working as the goal was missing proper header data, causing the
rotation shim to give back control to the main controller.