mppi parameters_handler: Improve verbose handling (#4704) (#4711)
* mppi parameters_handler: Improve verbose handling (#4704)

The "verbose" parameter of the parameters_handler is
a special case that needs registration before the
dynamic parameter handler callback is registered.

In verbose mode make the parameter handler info/warn/debug
messages more expressive.

Signed-off-by: Mike Wake <michael.wake@aosgrp.com.au>

* Address review comments. (#4704)

* remove comments.
* Use RCLCPP_DEBUG instead of INFO for low level messages.
* Add test for trying to access parameters that are not declared.

Signed-off-by: Mike Wake <michael.wake@aosgrp.com.au>

* mppi parameters_handler: Improve static/dynamic/not defined logging (#4704)

Attempts to change undefined parameters will not be successful
and will log an error.

Attempts to change static parameters will be ignored, a debug
message is logged if a change in parameters is attempted.

Signed-off-by: Mike Wake <michael.wake@aosgrp.com.au>

* mppi parameters_handler: populate SetParametersResult (#4704)

Provide a mechanism to populate an rcl_interfaces::msg::SetParametersResult
with the reasons for unsuccessful parameter setting, so that it may be
propogated back to a set parameter client.

The mechanism provides feedback when attempting to set undefined
parameters, static parameters and could be used to validate dynamic
parameters and provide a reason for rejection.

NOTE: This changes public interface of ParametersHandler class.
s/setDynamicParamCallback/setParamCallback
s/addDynamicParamCallback/addParamCallback which takes
a callback function that is able to populate a
rcl_interfaces::msg::SetParametersResult.

In order to indicate an unsuccessful parameter change and the
reason, callback functions should append a "\n"
to the reason before appending to if it is not empty.

Signed-off-by: Mike Wake <michael.wake@aosgrp.com.au>

* mppi parameters_handler: fix reason handling and improve tests

Signed-off-by: Mike Wake <michael.wake@aosgrp.com.au>

---------

Signed-off-by: Mike Wake <michael.wake@aosgrp.com.au>
Signed-off-by: Jakubach <jakubach@gmail.com>