Add grid value interpretation for inscribed inflated obstacles (#5781)
* nav2_costmap_2d: static_layer: add grid value interpretation for inscribed inflated obstacles

Until now, the special OccupancyGrid value "99" attributed to the
inscribed inflated obstacles was mapped to 251 on conversion to
Costmap2D, using the default linear relation. This is incorrect since a
special value exists in the Costmap2D for inscribed obstacles: 253.

This commit makes sure that the correct value is used in case of
inflated obstacle.

Signed-off-by: Dylan De Coeyer <dylan.decoeyer@quimesis.be>

* nav2_costmap_2d: fix unit tests following change

Signed-off-by: Dylan De Coeyer <dylan.decoeyer@quimesis.be>

* Fix tests

Signed-off-by: mini-1235 <mauricepurnawan@gmail.com>

* Fix tests

Signed-off-by: mini-1235 <mauricepurnawan@gmail.com>

* Lint

Signed-off-by: mini-1235 <mauricepurnawan@gmail.com>

---------

Signed-off-by: Dylan De Coeyer <dylan.decoeyer@quimesis.be>
Signed-off-by: mini-1235 <mauricepurnawan@gmail.com>
Co-authored-by: Dylan De Coeyer <dylan.decoeyer@quimesis.be>